Fix issue of dynamically updating regexes
This fixes an issue where a user could have added weasel words or
passive voice verbs and the regexes used would not have been updated to
handle the additional words.

They are now dynamic in nature and are updated upon entry into the minor
mode. This means any changes to the lists require a restart of
Writegood-mode.